Remove some bulky runtime deps
Apache and Coro::Select are only used by specific DBDs provided in this
package, and aren't required for *all* uses of DBI. Since they pull in a
lot of additional dependencies recursively, we'll leave them out here.

Also loosening up test requirements as there seem to be issues with the
upstream test suite.